add copy functionality to component bank codes
